starting the preview pipeline
